bugfix: fix the "let arg = args;" on each InstructionData node match, but this might have problems. see note for details.
earlier it was done using:
- let args_<global counter> = args; for binary, intCompare node
- let args_<global counter> = arg; for binaryImm, intcompareImm node

Since each such instruction data node knows the arg_name from
pctable that we really constructed while processing lhs (processlhs module)
and taken foward in further modules. So, RHS arg names were generated
correctly. But, the arg names in LHS were not mapped properly.

The idea is simple why not use the arg_name of the node itself
because this is literally holding the entire map of indices
of each argument and their names on both LHS and RHS.

However, there might be an issue of redefinition of args somewhere
when same arg_names are assigned to multiple nested instruction data nodes
like:
%0 = var
%1 = add %0, %0
%2 = mul %0, %0
infer %2
~>
..