fix the RHS code generation
- W=1 case in RHS for both insert and replace instructions. For W=1, it shpould be bconst. But for other Widths, it can
be iconst.
The integer value/literal part in immediate instructions or iconst needs
a proper type cast "1056565_u64 as i64"
Yet fox fix that any other width repkacement instruction on RHS can be any non-iconst inst as well. But that was only
one case.
particularly taking about cases here when on RHS there is no insertion of insts
but only single replace and that too replacement of LHS by one of the insts already defined in LHS. This would require
inst_results() and alias() stuff in code generation.